backend/pyproject.toml
[tool.poetry]
name = "suckyear-bot"
version = "0.1.0"
description = ""
authors = ["MikhailTo <toshkin.mikhail@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "api" },
    { include = "bot" },
    { include = "shared" }
]

[tool.poetry.scripts]
api = "api.main:run"
dbinit = "shared.database.commands:init"
makemigrations = "shared.database.commands:makemigrations"
migrate = "shared.database.commands:migrate"
rollback = "shared.database.commands:rollback"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
fluent-runtime = "^0.4.0"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
